backend/api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
torch>=2.2.0
transformers>=4.38.2
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-dotenv>=1.0.1
numpy>=1.26.4
pydantic>=2.6.3
google-generativeai>=0.3.0